UP-4993:  Refactor IPersonManager implementations slightly to avoid unnecessary bean tweaking to enable RemoteUserPersonManager
