+++ uPortal-webapp/src/main/resources/properties/contexts/userContext.xml
@@ -28,18 +28,18 @@
            http://www.springframework.org/schema/util http://www.springframework.org/schema/util/spring-util-3.1.xsd">
 
     <!--
-     | The concrete IPersonManager implementation to be used
+     | For uPortal 5, the various flavors of PersonManager were combined into a single class to
+     | avoid unnecessary bean tweaking on the part of deployers.  At some point in the future where
+     | we remove ExtendedPersonManager, it might be a good idea to remove this <bean> element and
+     | annotate RemoteUserPersonManager with @Component (or some such).
      +-->
-   	<bean id="personManager" class="org.apereo.portal.security.provider.SimplePersonManager" />
-    <!--
-     | Use this bean for Shibboleth authentication
-     +-->
-    <!--
     <bean id="personManager" class="org.apereo.portal.security.provider.RemoteUserPersonManager" />
-    -->
-   	<!--
-   	 | Use this bean to extend guest user properties with attributes from the request set by the requestAttributeSourceFilter
-   	 +-->
+
+    <!--
+     | The ExtendedPersonManager, however, still requires bean tweaking, if you're actually going to
+     | use it.  As far as we know, no one is using it.  The ExtendedPersonManager class has been
+     | deprecated.  To enable it in uPortal 5, you'll need to overlay this file.
+     +-->
    	<!--
    	<bean id="personManager" class="org.apereo.portal.security.provider.ExtendedPersonManager" >
    		<property name="merger">
@@ -59,7 +59,7 @@
         <property name="subMappers">
             <util:list>
                 <ref bean="sessionAttributeProfileMapper" />
-                
+
                 <!-- If the user's server dictates to use a profile, apply it -->
                 <!-- <ref bean="serverProfileMapper" />  -->
 
@@ -95,10 +95,10 @@
             </util:list>
         </property>
     </bean>
-    
+
     <!-- Example of how to use serverProfileMapper.  This puts users on profile=default
     for any server that matches the regex (everything not *.wisconsin.edu* for example -->
-    <!-- 
+    <!--
     <bean id="serverProfileMapper" class="org.apereo.portal.layout.profile.ServerProfileMapperImpl"
         p:serverRegex="^((?!.\\.wisconsin\\.edu).)*$"
         p:profile="default" />
